Add INIT-OPTIONS
Allows us to specify some initialization code *after* we have a GL context.

Also included is a change to our main loop. It allows us to more easily recover
from errors. For example, if have an undefined function in our main loop, when
the condition is raised we will be able to "Continue". However, so long as the
function is undefined the condition will repeatedly be raised. We can however
simply define the function, and afterwards "Continue". Without
LIVESUPPORT:CONTINUABLE we wouldn't be able to so.