Remove automatic deallocation of GL objects
-- How exiting the main loop used to work --

After we launch our render once

  CL-USER> (start-texture)

and then exit, we free all resources: every object in *vaos*, along with our
main STORE instance (*g*) is set to nil, allowing them to be GC'd. Meaning, when
we try to simply start it again (no recompiling), our GL objects used for the
render aren't stored in *vaos* any longer.

-- Now --

That *is* supposed to be what we want, however, in the context of interactive
development of shader programs, it's really just getting in our way. If this was
a project to write games, it'd be different. The better solution, for now, is to
have a manual function to free up resources when needed.

Hence, the removal of FREE from the exit of our main loop, and the added TODO.